+++ umask.go
@@ -3,6 +3,14 @@
 
 package compat
 
+import (
+	"errors"
+)
+
+const initialUmask = 0o022
+
+var ErrUmaskChanged = errors.New("process umask changed outside compat")
+
 func init() {
 	_ = initUmask()
 }